+++ kartoteka/__init__.py
@@ -1,0 +1,3 @@
+from .ui import CardEditorApp
+
+__all__ = ["CardEditorApp"]
