+++ tests/test_next_free_location.py
@@ -6,8 +6,8 @@ from unittest.mock import MagicMock
 
 sys.modules.setdefault("customtkinter", MagicMock())
 sys.path.append(str(Path(__file__).resolve().parents[1]))
-import main
-importlib.reload(main)
+import kartoteka.ui as ui
+importlib.reload(ui)
 
 
 def test_next_free_location_sequential():
@@ -18,10 +18,10 @@ def test_next_free_location_sequential():
             {"warehouse_code": "K01R1P0003"},
         ],
     )
-    dummy.generate_location = lambda idx: main.CardEditorApp.generate_location(dummy, idx)
-    first = main.CardEditorApp.next_free_location(dummy)
+    dummy.generate_location = lambda idx: ui.CardEditorApp.generate_location(dummy, idx)
+    first = ui.CardEditorApp.next_free_location(dummy)
     assert first == "K01R1P0002"
     dummy.output_data.append({"warehouse_code": first})
-    second = main.CardEditorApp.next_free_location(dummy)
+    second = ui.CardEditorApp.next_free_location(dummy)
     assert second == "K01R1P0004"
